added reload action
